+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] 2026-01-11
+
+### Updated
+- **Enum export**: Backed enums now export their value, unit enums export their name
+
 ## [1.2.0] - 2025-12-14
 
 ### Added